refactor : 필요없는 파일 제거(redisConfiguration없어도 정상동작함), 클래스명 변경(여러 서비스를 이용하는 걸로 봐서는 Facade가 적절한 네임으로 보임) #80
