--- src/main/kotlin/gomushin/backend/core/configuration/redis/RedisConfiguration.kt
@@ -1,14 +1,0 @@
-package gomushin.backend.core.configuration.redis
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.redis.connection.RedisConnectionFactory
-import org.springframework.data.redis.core.StringRedisTemplate
-
-@Configuration
-class RedisConfiguration {
-    @Bean
-    fun redisTemplate(redisConnectionFactory: RedisConnectionFactory): StringRedisTemplate {
-        return StringRedisTemplate(redisConnectionFactory)
-    }
-}
